requirements.txt

sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic[email]==2.11.3
python-dotenv==1.1.0
python-json-logger==2.0.7
loguru==0.7.2
fastapi==0.115.12
uvicorn==0.23.2
httpx==0.25.0

python-consul==1.1.0
prometheus_client==0.21.1
prometheus_fastapi_instrumentator==7.1.0
circuitbreaker==2.1.3

pytest==7.4.0
colorama==0.4.6
schedule==1.2.2

psutil==5.9.5
openai==1.76.0
langsmith==0.3.33
langchain==0.3.24
langchain-openai==0.3.14
langchain-core==0.3.55
langchain_groq==0.3.2
agno==1.4.1